py-tumblr/requirements.txt
python-tumblpy
Flask
gunicorn
